// types/MenteeType.ts
export interface MenteeType {
  _id: string;
  name: string;
  email: string;
  phone: string;
  whatsapp: string;
  ipAddress: string;
  ipDetails: any;
  college: string;
  year: number;
  github: string;
  linkedIn: string;
  isFirstTime: boolean;
  question1: string;
  answer1: string;
  isBanned: boolean;
}